[Shared/Models/Classes] Moved `PollOption` to its own file
