+++ Shared/Models/Classes/Poll.swift
@@ -7,29 +7,6 @@
 
 import Foundation
 
-/// A class representation of a poll option.
-public class PollOption: Codable, Identifiable {
-
-    // MARK: - STORED PROPERTIES
-
-    /// The ID for this poll option.
-    // swiftlint:disable:next identifier_name
-    public let id = UUID()
-
-    /// The title for this poll option.
-    public let title: String
-
-    /// The number of votes for this option.
-    public let votesCount: Int?
-
-    // MARK: - COMPUTED PROPERTIES
-
-    private enum CodingKeys: String, CodingKey {
-        case title
-        case votesCount = "votes_count"
-    }
-}
-
 /// A class representation of a poll.
 public class Poll: Codable, Identifiable {
